Updates presentation of job meta information
This commit updates the presentation of job meta information such as salary, location and seniority. The idea behind this being that too much information in the emoji detail row essentially renders it useless and that salary, location and seniority information is much more important to people than when the job was originally posted.